Tested the query on the database.
The new SQL query works even better. Thanks to Daan for the suggestion. Due to the limiting on the amount of times a query is executed this should be faster and less demanding on the server. However I do not know for certain if it is a weighted average or not while in the old one it was certain the average was not weighted.