Allowed the DatabaseConnector to be passed as a runtime argument. The test now also has a mockito test in it that checks if the non database stuff of the class is working properly.
However the pom file does not contain mockito yet in this branch, but since it already does elsewhere we let it slide here to prevent unnecessary conflicts.

Due to the database branch updates the test code is not ready to be executed at this moment in time. This should come after both this branch and the database branch are merged in the master branch. Furthermore the query to be executed now requested through the getQuery() method. This makes extension way easier if the need ever seems to arise.