Created a CheckStyle file specifically for maven. This is because the checkstyle-maven-plugin is only at CheckStyle version 5.8.
The maven CheckStyle file contains less tests than the other (read all non existing tests in version 5.8 have been removed) and as such presents a clear yet not entirely true view on the code. However only 3 tests where incompatible so the result is very reliable. The documentation of the CheckStyle file is going to be updated on this in the docs repository.

Changed the CheckStyle file so that max line length is set to 100. I also fixed the bug that all packages should have a package-info.java since this was a bit extreme and caused standard errors since the test and code were in the same package in different folders and it is an error to make 2 package-infos. Doing this is actually good practice and makes everything look good and structured.