Increased the amount of testing done on the Shingle class.
The Shingle class is now more tested by it's own test case. Even the methods like unique and union now are tested. it now contains of 34 different tests.

Added several methods to the Shingle class that are useful when testing/debugging and/or when we want to store/compare shingles. It now has an equals and hashCode method as well as a toString.