Added uncommited files and minor fixes.
Because of the new routing there now is a workaround issue #16. By
reloading the widget the page will not re-render and delete the
accesstoken. The old route still exists just for sake of utility but
will NOT keep you logged in when you do choose to use it. Since the
standard way of previewing has changed this fixes #16.