Fixed the failing tests and altered the way in which the comment is searched for the words.
Comments are now no longer searched with contains but are split in words and those words are then processed one by one. This increases efficiency because contains is an O(nm) done x times, with x being the amount of words worth any points, while the new method is O(wb) done once. In general a speedup.