Removed the completely unnecessary text file. Please do not recommit this file.
This file is generated by one of the tests. This test includes the removing of this file. If this test fails on your local machine and this file is not removed, do remove it yourself, but do not push it to GitHub. This file does not belong in the project but is automatically generated and removed by one of the test cases.