Fixed the bug that you cannot accidentally hack the program and get a StackOverflowError. You now get a IllegalStateException when it is obviously going in that direction.
I also added test cases that test that this is indeed the case.

threshold difference has been increased for quicker searching of the pieces that are requested. This makes the search a little less reliable but it makes it a lot faster, since now only 1 step has to be done while earlier you needed three. The decrease in reliability is because now the interfall is five times as large so you mix together more pieces into the category. However since you the jaccard distance increases very quickly it is very unlikely that this matters at all, especially on the important points were the threshold is still a large value.