Removed merge errors and duplicate code.
Duplication in the seekers is removed, since the seek methods where all exactly the same now that the calculation is done elsewhere (due to the Decorator pattern). This makes it easier to change it and allows less tests to make all the paths covered. It also reduces typing in the future.